Add routes to nodejs
